% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{install_maven}
\alias{install_maven}
\title{Install Maven}
\usage{
install_maven(dir = NULL)
}
\arguments{
\item{dir}{(Optional) Directory to install maven in.
Defaults to \code{maven/} under user's home directory.}
}
\description{
This function installs Apache Maven.
}
