% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{install_mleap}
\alias{install_mleap}
\title{Install MLeap runtime}
\usage{
install_mleap(dir = NULL, version = NULL)
}
\arguments{
\item{dir}{(Optional) Directory to save the jars}

\item{version}{Version of MLeap to install, defaults to the latest version tested with this package.}
}
\description{
Install MLeap runtime
}
