% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bundle.R
\name{ml_read_bundle}
\alias{ml_read_bundle}
\title{Loads an MLeap bundle into Spark}
\usage{
ml_read_bundle(path)
}
\arguments{
\item{path}{Path to the exported bundle zip file.}
}
\value{
An Spark ML Pipeline model object.
}
\description{
Loads an MLeap bundle into Spark
}
