% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleap.R
\name{ml_write_bundle}
\alias{ml_write_bundle}
\title{Export a Spark pipeline for serving}
\usage{
ml_write_bundle(x, dataset, path, overwrite = FALSE)
}
\arguments{
\item{x}{A Spark pipeline model object.}

\item{dataset}{A Spark DataFrame with the schema of the transformed DataFrame.}

\item{path}{Where to save the bundle.}

\item{overwrite}{Whether to overwrite an existing file, defaults to \code{FALSE}.}
}
\description{
This functions serializes a Spark pipeline model into an MLeap bundle.
}
