% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{mleap_defaults}
\alias{mleap_defaults}
\title{Gets the processed defaults used during the MLeap session}
\usage{
mleap_defaults(var = NULL)
}
\arguments{
\item{var}{Variable to extract from defaults}
}
\description{
Gets the processed defaults used during the MLeap session
}
