% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleap_dep_versions.R
\name{mleap_dep_versions_table}
\alias{mleap_dep_versions_table}
\title{A table containing the combinations of Spark, Scala and MLeap versions to use}
\usage{
mleap_dep_versions_table()
}
\description{
A table containing the combinations of Spark, Scala and MLeap versions to use
}
