% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleap_dep_versions.R
\name{mleap_dependency_versions}
\alias{mleap_dependency_versions}
\title{A table containing the combinations of Spark, Scala and MLeap versions to use}
\usage{
mleap_dependency_versions()
}
\description{
A table containing the combinations of Spark, Scala and MLeap versions to use
}
