% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{mleap_installed_versions}
\alias{mleap_installed_versions}
\title{Find existing MLeap installations}
\usage{
mleap_installed_versions()
}
\value{
A data frame of MLeap Runtime installation versions and
  their locations.
}
\description{
Find existing MLeap installations
}
