% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleap.R
\name{mleap_load_bundle}
\alias{mleap_load_bundle}
\title{Loads an MLeap bundle}
\usage{
mleap_load_bundle(path)
}
\arguments{
\item{path}{Path to the exported bundle zip file.}
}
\value{
An MLeap model object.
}
\description{
Loads an MLeap bundle
}
