% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleap.R
\name{mleap_model_schema}
\alias{mleap_model_schema}
\title{MLeap model schema}
\usage{
mleap_model_schema(x)
}
\arguments{
\item{x}{An MLeap model object.}
}
\value{
A data frame of the model schema.
}
\description{
Returns the schema of an MLeap transformer.
}
