% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{mleap_transform}
\alias{mleap_transform}
\title{Transform data using an MLeap model}
\usage{
mleap_transform(model, data)
}
\arguments{
\item{model}{An MLeap model object, obtained by \code{mleap_load_bundle()}.}

\item{data}{An R data frame.}
}
\value{
A transformed data frame.
}
\description{
This functions transforms an R data frame using an MLeap model.
}
\seealso{
[mleap_load_bundle()]
}
