% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{install_maven}
\alias{install_maven}
\title{Install Maven}
\usage{
install_maven(dir = NULL, version = NULL)
}
\arguments{
\item{dir}{(Optional) Directory to install maven in.
Defaults to \code{maven/} under user's home directory.}

\item{version}{Version of Maven to install, defaults to the latest version tested with this package.}
}
\description{
This function installs Apache Maven.
}
\examples{
\dontrun{
install_maven()
}

}
