% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{install_mleap}
\alias{install_mleap}
\title{Install MLeap runtime}
\usage{
install_mleap(dir = NULL, restart_session = TRUE)
}
\arguments{
\item{dir}{(Optional) Directory to save the jars}

\item{restart_session}{Whether to restart R session after installation}
}
\description{
Install MLeap runtime
}
