% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mleap.R
\name{install_mleap}
\alias{install_mleap}
\title{Install MLeap runtime}
\usage{
install_mleap(
  mleap_version = mleap_defaults("version"),
  force = FALSE,
  scala_version = "auto",
  dir = mleap_defaults("base_folder"),
  use_temp_cache = TRUE
)
}
\arguments{
\item{mleap_version}{Version of MLeap to install.  Defaults to what is processed by `mleap_session()`.}

\item{force}{Force the installation. Defaults to \code{FALSE}.}

\item{scala_version}{Version of Scala to match to the MLeap version. It defaults to 'auto'.
If 'auto', the installation process will determine the most recent version of Scala
to use available for the requested MLeap version. Other possible values are '2.10', 
'2.11', and '2.12'.}

\item{dir}{Directory to save the jars. Defaults to what is processed by `mleap_session()`.}

\item{use_temp_cache}{Whether to use a temporary Maven cache directory for downloading.
Setting this to \code{TRUE} prevents Maven from creating a persistent \code{.m2/} directory.
Defaults to \code{TRUE}.}
}
\description{
Install MLeap runtime
}
\examples{
\dontrun{
install_mleap()
}
}
