% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{install_mleap}
\alias{install_mleap}
\title{Install MLeap runtime}
\usage{
install_mleap(dir = NULL, version = NULL, use_temp_cache = TRUE)
}
\arguments{
\item{dir}{(Optional) Directory to save the jars}

\item{version}{Version of MLeap to install, defaults to the latest version tested with this package.}

\item{use_temp_cache}{Whether to use a temporary Maven cache directory for downloading.
Setting this to \code{TRUE} prevents Maven from creating a persistent \code{.m2/} directory.
Defaults to \code{TRUE}.}
}
\description{
Install MLeap runtime
}
\examples{
\dontrun{
install_mleap()
}
}
